refactor(app): move the home screen state into ainb-app
HomeScreenV2State and the state it is built from move to ainb-app: the
sidebar's items, selection and width clamp, the welcome panel's scroll
state and the mascot animation. Their renderers stay in ainb-core.

The state's copy of the last drawn sidebar rectangle, which a click is
hit-tested against, becomes an Area; publish_after_draw converts the
renderer's Rect when it copies it across. The row hit-test
(item_index_at) takes that Area and moves with the sidebar state. ainb-app
picks up arboard for the welcome panel's copy action.
